Rearrange namespaces and thus directory layout
