--- lib/data_mapper/mapper/relationship/one_to_one.rb
@@ -1,29 +1,0 @@
-module DataMapper
-  class Mapper
-
-    # Attribute
-    #
-    # @api private
-    class Relationship
-
-      class OneToOne < Relationship
-
-        private
-
-        def default_source_key
-          :id
-        end
-
-        def default_target_key
-          foreign_key_name
-        end
-
-        # @api private
-        def relationship_builder
-          Builder::OneToOne
-        end
-      end # class OneToOne
-
-    end # class Relationship
-  end # class Mapper
-end # module DataMapper
