--- lib/data_mapper/mapper/relationship/options/many_to_one.rb
@@ -1,23 +1,0 @@
-module DataMapper
-  class Mapper
-    class Relationship
-      class Options
-
-        class ManyToOne < self
-
-          def type
-            Relationship::ManyToOne
-          end
-
-          def default_source_key
-            foreign_key_name(target_model.name)
-          end
-
-          def default_target_key
-            :id
-          end
-        end # class ManyToOne
-      end # class Options
-    end # class Relationship
-  end # class Mapper
-end # module DataMapper
