--- lib/data_mapper/mapper/relationship/options/one_to_one.rb
@@ -1,23 +1,0 @@
-module DataMapper
-  class Mapper
-    class Relationship
-      class Options
-
-        class OneToOne < self
-
-          def type
-            Relationship::OneToOne
-          end
-
-          def default_source_key
-            :id
-          end
-
-          def default_target_key
-            foreign_key_name(source_model.name)
-          end
-        end # class OneToOne
-      end # class Options
-    end # class Relationship
-  end # class Mapper
-end # module DataMapper
