+++ lib/data_mapper/query.rb
