+++ lib/data_mapper/relationship/builder/one_to_one.rb
@@ -1,0 +1,14 @@
+module DataMapper
+  class Relationship
+    class Builder
+
+      class OneToOne < self
+        private
+
+        def fields
+          super.merge(source_key => target_key)
+        end
+      end # class OneToOne
+    end # class Builder
+  end # class Relationship
+end # module DataMapper
