+++ lib/dm-mapper.rb
@@ -72,40 +72,41 @@ end # module DataMapper
 
 require 'veritas'
 require 'veritas-optimizer'
+require 'virtus/support/descendants_tracker'
 
 require 'data_mapper/support/utils'
+require 'data_mapper/support/inflector/inflections'
+require 'data_mapper/support/inflector/methods'
+require 'data_mapper/support/inflections'
 
+require 'data_mapper/relation_registry'
+require 'data_mapper/mapper_registry'
+require 'data_mapper/relationship/dsl'
+
+require 'data_mapper/mapper/relationship_set'
 require 'data_mapper/mapper/attribute'
 require 'data_mapper/mapper/attribute/primitive'
 require 'data_mapper/mapper/attribute/mapper'
 require 'data_mapper/mapper/attribute/collection'
 require 'data_mapper/mapper/attribute_set'
-
-require 'data_mapper/mapper/relationship'
-require 'data_mapper/mapper/relationship/options'
-require 'data_mapper/mapper/relationship/options/one_to_one'
-require 'data_mapper/mapper/relationship/options/one_to_many'
-require 'data_mapper/mapper/relationship/options/many_to_one'
-require 'data_mapper/mapper/relationship/options/many_to_many'
-require 'data_mapper/mapper/relationship/one_to_one'
-require 'data_mapper/mapper/relationship/one_to_many'
-require 'data_mapper/mapper/relationship/many_to_one'
-require 'data_mapper/mapper/relationship/many_to_many'
-require 'data_mapper/mapper/relationship/builder'
-require 'data_mapper/mapper/relationship/builder/collection_behavior'
-require 'data_mapper/mapper/relationship/builder/one_to_one'
-require 'data_mapper/mapper/relationship/builder/one_to_many'
-require 'data_mapper/mapper/relationship/builder/many_to_one'
-require 'data_mapper/mapper/relationship/builder/many_to_many'
-require 'data_mapper/mapper/relationship_set'
-
-require 'data_mapper/mapper/mapper_registry'
-require 'data_mapper/mapper/relation_registry'
 require 'data_mapper/mapper'
 require 'data_mapper/mapper/veritas_mapper'
 
-require 'data_mapper/mapper/query'
+require 'data_mapper/relationship'
+require 'data_mapper/relationship/options'
+require 'data_mapper/relationship/options/one_to_one'
+require 'data_mapper/relationship/options/one_to_many'
+require 'data_mapper/relationship/options/many_to_one'
+require 'data_mapper/relationship/options/many_to_many'
+require 'data_mapper/relationship/one_to_one'
+require 'data_mapper/relationship/one_to_many'
+require 'data_mapper/relationship/many_to_one'
+require 'data_mapper/relationship/many_to_many'
+require 'data_mapper/relationship/builder'
+require 'data_mapper/relationship/builder/collection_behavior'
+require 'data_mapper/relationship/builder/one_to_one'
+require 'data_mapper/relationship/builder/one_to_many'
+require 'data_mapper/relationship/builder/many_to_one'
+require 'data_mapper/relationship/builder/many_to_many'
 
-require 'data_mapper/support/inflector/inflections'
-require 'data_mapper/support/inflector/methods'
-require 'data_mapper/support/inflections'
+require 'data_mapper/query'
