+++ spec/unit/data_mapper/mapper/class_methods/mapper_registry_spec.rb
@@ -2,6 +2,6 @@ require 'spec_helper'
 
 describe DataMapper::Mapper, '.mapper_registry' do
   it "returns mapper registry instance" do
-    described_class.mapper_registry.should be_instance_of(described_class::MapperRegistry)
+    described_class.mapper_registry.should be_instance_of(DataMapper::MapperRegistry)
   end
 end
