+++ spec/unit/data_mapper/mapper/class_methods/relation_registry_spec.rb
@@ -2,6 +2,6 @@ require 'spec_helper'
 
 describe DataMapper::Mapper, '.relation_registry' do
   it "returns mapper registry instance" do
-    described_class.relation_registry.should be_instance_of(described_class::RelationRegistry)
+    described_class.relation_registry.should be_instance_of(DataMapper::RelationRegistry)
   end
 end
