--- spec/unit/data_mapper/mapper/mapper_registry/append_spec.rb
@@ -1,12 +1,0 @@
-require 'spec_helper'
-
-describe DataMapper::Mapper::MapperRegistry, '#<<' do
-  let(:model)    { mock('model') }
-  let(:mapper)   { mock_mapper(model).new }
-  let(:registry) { described_class.new }
-
-  it "adds a new mapper to the registry" do
-    registry << mapper
-    registry[model].should be(mapper)
-  end
-end
