--- spec/unit/data_mapper/mapper/relation_registry/relation_assignment_spec.rb
@@ -1,12 +1,0 @@
-require 'spec_helper'
-
-describe DataMapper::Mapper::RelationRegistry, '#[]=' do
-  let(:name)     { :users }
-  let(:relation) { mock('relation', :name => name) }
-  let(:registry) { described_class.new }
-
-  it 'adds relation to the registry' do
-    registry[name] = relation
-    registry[name].should be(relation)
-  end
-end
