--- spec/unit/data_mapper/mapper/relationship_dsl/belongs_to_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe DataMapper::Mapper::RelationshipDsl, '.belongs_to' do
-  it 'creates ManyToOne relationship' do
-    pending
-  end
-end
