--- spec/unit/data_mapper/mapper/relationship_dsl/has_many_spec.rb
@@ -1,11 +1,0 @@
-require 'spec_helper'
-
-describe DataMapper::Mapper::RelationshipDsl, '.has_many' do
-  it "creates OneToMany relationship" do
-    pending
-  end
-
-  it "creates ManyToOne relationship" do
-    pending
-  end
-end
