--- spec/unit/data_mapper/mapper/relationship_dsl/has_one_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe DataMapper::Mapper::RelationshipDsl, '.has' do
-  it 'creates OneToOne relationship' do
-    pending
-  end
-end
