+++ spec/unit/data_mapper/relation_registry/append_spec.rb
@@ -1,0 +1,12 @@
+require 'spec_helper'
+
+describe DataMapper::RelationRegistry, '#<<' do
+  let(:name)     { :users }
+  let(:relation) { mock('relation', :name => name) }
+  let(:registry) { described_class.new }
+
+  it "adds relation to the registry" do
+    registry << relation
+    registry[name].should be(relation)
+  end
+end
