+++ spec/unit/data_mapper/relationship/dsl/belongs_to_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe DataMapper::Relationship::Dsl, '.belongs_to' do
+  it 'creates ManyToOne relationship' do
+    pending
+  end
+end
