+++ spec/unit/data_mapper/relationship/dsl/has_many_spec.rb
@@ -1,0 +1,11 @@
+require 'spec_helper'
+
+describe DataMapper::Relationship::Dsl, '.has_many' do
+  it "creates OneToMany relationship" do
+    pending
+  end
+
+  it "creates ManyToOne relationship" do
+    pending
+  end
+end
