+++ spec/unit/data_mapper/relationship/dsl/has_one_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe DataMapper::Relationship::Dsl, '.has' do
+  it 'creates OneToOne relationship' do
+    pending
+  end
+end
